backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

langgraph==0.2.60
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-google-genai==2.0.8
langchain-chroma==0.1.4
langchain-text-splitters==0.3.4

chromadb==0.5.23
google-generativeai==0.8.3

ddgs>=9.10.0
duckduckgo-search==6.4.2
pypdf==5.1.0
docx2txt==0.8
unstructured==0.16.11
